renamed files from master-teachers to master-teachers-table
